apps/replikate/requirements.txt
kopf
pykube-ng
ptvsd
pydevd
pyyaml
Jinja2
# watchdog
